- working on popover contents - fix bug when deleting a study and there're no files to delete. this issue throws an error. fixed it by filter and checking array - redesign json format for feedback and scenes data - fix several bugs and issues for study creation forms

+++ includes/delete-study.php
@@ -119,7 +119,10 @@ if (isset($_SESSION['user_id'], $_POST['studyId'])) {
                         echo json_encode(array('status' => 'deleteError'));
                         exit();
                     } else {
-                        deleteFiles($target_dir, $deleteFiles);
+                        $deleteFiles = array_filter($deleteFiles);
+                        if (!empty($deleteFiles)) {
+                            deleteFiles($target_dir, $deleteFiles);
+                        }
                         echo json_encode(array('status' => 'success', 'deletedFiles' => $deleteFiles));
                         exit();
                     }
